Move MempoolStore to a submodule
